requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
anthropic==0.25.0
openai==1.12.0
youtube-transcript-api==0.6.2
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
pillow==10.2.0
gspread==6.0.2
tqdm==4.66.2
tenacity==8.2.3
backoff==2.2.1
sentence-transformers==2.6.1
numpy==1.26.4
scikit-learn==1.4.1
